libbpf-cargo: return stderr from clang
Update `build_and_generate` to return the `Vec<u8>` that is already collected
from the clang output's stderr. This allows consumers to decide what to do with
that information.

The primary consumer is `sched-ext/scx` which processes the `Vec` with
`String::from_utf8_lossy` and outputs each line with
`println!("cargo:warning={}", l)` which shows the warnings quite nicely in the
Cargo output when there isn't an error.

We could alternatively process the String on our side to get a more explicit
output, but this requires reallocating the data. Given consumers won't benefit
from this by default, the extra allocation seems pointless.

Test plan:
- Pointed that Cargo.lock at this repo and built.