libbpf-cargo: return stderr from clang
Update `build_and_generate` to return a `CompilationOutput` owning the stderr
of the compiler process. This gives consumers access to the stderr to output as
they see fit.

Provides a `stderr` method that calls `String::from_utf8_lossy`. The primary
consumer is `sched-ext/scx` which outputs each line with
`println!("cargo:warning={}", l)`, showing the warnings quite nicely in the
Cargo output when there isn't an error.

Test plan:
- Pointed that Cargo.lock at this repo and built.